refactoring Instructor class
